Break up C files, consolidate H files
