+++ source/http.h
@@ -1,5 +1,5 @@
-#ifndef AWS_CRT_PYTHON_HTTP_STREAM_H
-#define AWS_CRT_PYTHON_HTTP_STREAM_H
+#ifndef AWS_CRT_PYTHON_HTTP_H
+#define AWS_CRT_PYTHON_HTTP_H
 /*
  * Copyright 2010-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
@@ -16,11 +16,22 @@
  */
 #include "module.h"
 
-struct aws_http_message;
+/**
+ * Close the connection if it's open.
+ */
+PyObject *aws_py_http_connection_close(PyObject *self, PyObject *args);
 
 /**
- * Initiates a request on connection.
+ * Returns True if connection is open and usable, False otherwise.
  */
+PyObject *aws_py_http_connection_is_open(PyObject *self, PyObject *args);
+
+/**
+ * Create a new connection. returns void. The on_setup callback will be invoked
+ * upon either success or failure of the connection.
+ */
+PyObject *aws_py_http_client_connection_new(PyObject *self, PyObject *args);
+
 PyObject *aws_py_http_client_stream_new(PyObject *self, PyObject *args);
 
 PyObject *aws_py_http_request_new(PyObject *self, PyObject *args);
@@ -28,7 +39,8 @@ PyObject *aws_py_http_request_new(PyObject *self, PyObject *args);
 /* Given a python object, return a pointer to its underlying native type.
  * If NULL is returned, a python error has been set */
 
+struct aws_http_connection *aws_py_get_http_connection(PyObject *connection);
 struct aws_http_stream *aws_py_get_http_stream(PyObject *stream);
 struct aws_http_message *aws_py_get_http_message(PyObject *http_message);
 
-#endif /* AWS_CRT_PYTHON_HTTP_STREAM_H */
+#endif /* AWS_CRT_PYTHON_HTTP_H */
