+++ source/http_connection.c
@@ -12,7 +12,7 @@
  * express or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-#include "http_connection.h"
+#include "http.h"
 
 #include "io.h"
